server_service/requirements.txt
django
polars
requests
bs4
pyyaml
